CommentStats수정

+++ src/main/java/com/example/blogex/domain/comment/dto/CommentStats.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CommentStats {
     private Long id;
-    private int cntLike;
+    private Long cntLike;
 }
